WIP: Support optimized  RasterizePoints map algebra function
 - for issue #146
 - Re-factor the existing rasterize vector with a base class of
   functionality to be shared with rasterize points.
